backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.1",
    "python-docx>=1.2.0",
    "uvicorn>=0.41.0",
]
